Added Guest Delete MessageBox

+++ Presentation Layer/6. FAQs.cs
@@ -162,7 +162,14 @@ namespace Phumla_Kamnandi.Presentation_Layer
             if (GuestView.SelectedRows.Count > 0)
             {
                 string SelectedGuest = GuestView.SelectedRows[0].Cells[0].Value.ToString();
-                guestController.DeleteGuest(SelectedGuest);
+                DialogResult DeleteConfirmation = MessageBox.Show(" Are you sure you want  to delete the guest with the following details " + SelectedGuest, "Confirm Deletetion ", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+
+                if (DeleteConfirmation == DialogResult.Yes)
+                {
+                    guestController.DeleteGuest(SelectedGuest);
+                    MessageBox.Show("Guest Deleted Successfully ", "Deletion Success", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                }
+                
 
                
             }
